workspace where the proc-macro is michie-macro
