+++ macro/README.md
@@ -1,0 +1,1 @@
+See the parent crate: <https://crates.io/crates/michie>
